Restructure into package
