+++ videocut/core/transcribe.py
